move a load of the options out from init into options.lua
